requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
requests
beautifulsoup4
pytesseract
Pillow
nltk
joblib